Deleting ellie_ocado file
